Shotgun: make snapshot of entire /var/log
Shootgun uses a fabric.api.get method to download some folder from the
remote host. Unfortunately, this method has a bug and fails when we
trying to download a folder that contains a symlink to another one.

So we need to patch fabric api to prevent us from such falls. The issue
with fabric api is that it recognize all symlinks as a file. As a result
we fails when trying to download a folder (pointed by a symlink)  as
a file.

The patch was proposed to the fabric master branch and can be reviewed
here:

    https://github.com/fabric/fabric/pull/1147

Closes-Bug: #1318514
Closes-Bug: #1323436

Change-Id: Ic2ae889337e5da15bbdf83338b8e8d5d913aeb63