Remove `local` flag
For now `shotgun` always executes on the `mcollective` docker
container. `local` flag was created to be able to execute a command on
`localhost` which, in the past, was master node. But for now it is
`mcollective` and as the result we get diagnostic snapshots from
`mcollective`. There is no any reason to have `local` flag for execution
a command on the `localhost` (it will be always `mcollective`
container). If we need to get diagnostic snapshots for some particular
container we can create command for such purpose. For example:

DUMP:
  dump:
    master:
      objects:
        - type: command
          command: `dockerctl shell mcollective rpm -qa`
          to_file: mcollective_installed_rpms.txt

Change-Id: Iac1d9ea37400e4d202a9bc673c5fa3d3d185701b
Partial-Bug: #1511730