Skip offline nodes
Apparently, shotgun never tested which nodes are online
prior to execute snapshot method of objects. Therefore, for offline
nodes shotgun will then consume a lot of time waiting for connection
timeouts to be happened on per object basis.

This patch totally rework how shotgun hanldes the objects:
* objects being processed on per host basis
* on NetworkError, all remaining objects on per host basis will be
  postponed for processing at the end of queue.
* if postponed objects still couldn't be proccessed, they will then
  be converted into offline objects.
* adds `Offline` object driver. Being snapshotted, this object just
  creates a file with friendly reminder that particular node was
  offline inside of snapshot target directory. So, one which's
  looking at snapshot archive will effectively realize that node
  was offline without deep diving into shotgun log file.

So, shotgun won't waste time or per object basis, as it was before.

Change-Id: Ib99476c7b67a04d4f472f3fa3803b3fb92d4fec4
Closes-Bug: #1397038