Dump shotgun log at the end of snapshot creation
Shotgun dumps /var/log/shotgun.log as normal part of the snapshot, but
it fills it in as the snapshot is created, which resulted in dumping
shotgun log split in half.

This patch makes shotgun dump it's own logs at the end of snapshot
creation process, just before compressing it.

Change-Id: Ie3e9737b2f430fd2f6ff3083b216f0ee62681c97
Closes-Bug: #1496842