Add report subcommand
This subcommand is to be used instead of version.yaml

Version.yaml used to provide short info
about what is installed on the master node.
We've switched to the package based
delivery approach and using build time artifact
(version.yaml) while packages could be easily
updated using yum command does not look rational.

This patch makes it possible to gather a short
report about what is installed on the master node
as well as about the content of some file in
a data driven manner.

One can run

shotgun report --config /path/to/report_config.yaml

/etc/shotgun/report.yaml is used by default as a config for
the report subcommand. It is to be provided by
shotgun rpm package.

Change-Id: I304a972c0c275d23ce2896a4d4e60870f9504fb4
Related-Bug: #1515517