Create a symlink to snapshot.
This change allows to define a target_symlink property in shotgun
config, which specifies a path to create a snapshot symlink. This
mechanism can be used to fix problems with insufficient disk space in
place that shotgun needs to create a dump (create a dump elsewhere and
link to it where we want to have it).

Change-Id: If459656772ef54c083fcdaabbd8d990f16b52db0
Related-bug: 1529182